Converted to use the new floating-point output routine.

+++ CommonCode/CommonLibrary/reportAnything.h
@@ -46,7 +46,7 @@
 void reportAnything(t_object *   xx,\
                     const char * routineName,\
                     t_symbol *   message,\
-                    const short  argc,\
+                    const long   argc,\
                     t_atom *     argv)
 
 REPORTANYTHING_HEADER;
